Parse files

+++ src/Reservoir.cpp
@@ -1,13 +1,13 @@
 #include "../header/Reservoir.h"
 
-Reservoir::Reservoir(std::string &name, std::string &municipality, std::string &code, uint32_t id, uint32_t maxDelivery):
+Reservoir::Reservoir(std::string &name, std::string &municipality, std::string &code, uint32_t id, double maxDelivery):
     name(std::move(name)), municipality(std::move(municipality)), Node(id, code), maxDelivery(maxDelivery) {}
 
 const std::string &Reservoir::getName() const {
     return name;
 }
 
-uint32_t Reservoir::getMaxDelivery() const {
+double Reservoir::getMaxDelivery() const {
     return maxDelivery;
 }
